# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: ComplyTime Platform API
  version: 0.1.0
  description: >
    REST surface served by `cmd/gateway`. Many `/api/*` routes register only when PostgreSQL
    is configured; `/api/config` is always available. After Google OAuth, the encrypted
    `studio_session` cookie authenticates `/api/*` requests (unless disabled). Optional
    `Authorization: Bearer <STUDIO_API_TOKEN>` bypasses the cookie for scripting.
    Security schemes below model bearer token and OAuth; cookie sessions are established via `/auth/*`.
servers:
  - url: http://localhost:8080
    description: Local gateway default (PORT=8080)
security:
  - bearerAuth: []
  - oauth2: [openid, email, profile]
...
paths:
...
  /api/posture:
    get:
      tags: [Posture]
      operationId: listPosture
      parameters:
        - name: start
          in: query
          description: RFC3339 or date-only (YYYY-MM-DD)
          schema:
            type: string
        - name: end
          in: query
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostureRow'
        '400':
          description: Invalid time range
  /api/requirements:
    get:
      tags: [Requirements]
      operationId: listRequirementMatrix
      parameters:
        - name: policy_id
          in: query
          required: true
          schema:
            type: string
        - name: audit_start
          in: query
          schema:
            type: string
        - name: audit_end
          in: query
          schema:
            type: string
        - name: classification
          in: query
          schema:
            type: string
        - name: control_family
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RequirementRow'
        '400':
          description: Invalid parameters
  /api/requirements/{id}/evidence:
    get:
      tags: [Requirements]
      operationId: listRequirementEvidence
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Requirement or control id
        - name: policy_id
          in: query
          required: true
          schema:
            type: string
        - name: audit_start
          in: query
          schema:
            type: string
        - name: audit_end
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RequirementEvidenceRow'
        '404':
          description: Requirement unknown for policy

  # Export endpoints (CSV, Excel, PDF) are planned but not yet implemented.
  /api/threats:
    get:
      tags: [ThreatsRisks]
      operationId: listThreats
      parameters:
        - name: catalog_id
          in: query
          schema:
            type: string
        - name: policy_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ThreatRow'
  /api/control-threats:
    get:
      tags: [ThreatsRisks]
      operationId: listControlThreats
      parameters:
        - name: catalog_id
          in: query
          schema:
            type: string
        - name: control_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ControlThreatRow'
  /api/risks:
    get:
      tags: [ThreatsRisks]
      operationId: listRisks
      parameters:
        - name: catalog_id
          in: query
          schema:
            type: string
        - name: policy_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RiskRow'
  /api/risks/severity:
    get:
      tags: [ThreatsRisks]
      operationId: listRiskSeverityByPolicy
      parameters:
        - name: policy_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RiskSeverityRow'
  /api/risk-threats:
    get:
      tags: [ThreatsRisks]
      operationId: listRiskThreats
      parameters:
        - name: catalog_id
          in: query
          schema:
            type: string
        - name: risk_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RiskThreatRow'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Shared secret from `STUDIO_API_TOKEN`; treated as admin-capable service identity when configured.
    oauth2:
      type: oauth2
      description: >
        Google OAuth 2.0 login flow used to mint the encrypted `studio_session` cookie.
        API requests authenticate via that cookie unless `bearerAuth` matches.
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/v2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            openid: OpenID Connect
            email: Email
            profile: Profile
...
  schemas:
...
    PostureRow:
      type: object
      properties:
        policy_id:
          type: string
        title:
          type: string
        version:
          type: string
        total_rows:
          type: integer
          minimum: 0
        passed_rows:
          type: integer
          minimum: 0
        failed_rows:
          type: integer
          minimum: 0
        other_rows:
          type: integer
          minimum: 0
        latest_at:
          type: string
        target_count:
          type: integer
          minimum: 0
        control_count:
          type: integer
          minimum: 0
        latest_evidence_at:
          type: string
        owner:
          type: string
    RequirementRow:
      type: object
      properties:
        catalog_id:
          type: string
        control_id:
          type: string
        control_title:
          type: string
        requirement_id:
          type: string
        requirement_text:
          type: string
        evidence_count:
          type: integer
          minimum: 0
        latest_evidence:
          type: string
        classification:
          type: string
    RequirementEvidenceRow:
      type: object
      properties:
        evidence_id:
          type: string
        target_id:
          type: string
        target_name:
          type: string
        rule_id:
          type: string
        eval_result:
          type: string
        classification:
          type: string
        assessed_at:
          type: string
        collected_at:
          type: string
        source_registry:
          type: string
...
    ThreatRow:
      type: object
      properties:
        catalog_id:
          type: string
        threat_id:
          type: string
        title:
          type: string
        description:
          type: string
        group_id:
          type: string
        policy_id:
          type: string
    ControlThreatRow:
      type: object
      properties:
        catalog_id:
          type: string
        control_id:
          type: string
        threat_reference_id:
          type: string
        threat_entry_id:
          type: string
    RiskRow:
      type: object
      properties:
        catalog_id:
          type: string
        risk_id:
          type: string
        title:
          type: string
        description:
          type: string
        severity:
          type: string
        group_id:
          type: string
        impact:
          type: string
        policy_id:
          type: string
    RiskThreatRow:
      type: object
      properties:
        catalog_id:
          type: string
        risk_id:
          type: string
        threat_reference_id:
          type: string
        threat_entry_id:
          type: string
    RiskSeverityRow:
      type: object
      properties:
        control_id:
          type: string
        max_severity:
          type: string
        risk_count:
          type: integer
          minimum: 0
...